Allow renaming collection in view playlist page.
