+++ ui/constants/action_types.js
@@ -195,7 +195,6 @@ export const COLLECTION_ITEMS_RESOLVE_STARTED = 'COLLECTION_ITEMS_RESOLVE_STARTE
 export const COLLECTION_ITEMS_RESOLVE_COMPLETED = 'COLLECTION_ITEMS_RESOLVE_COMPLETED';
 export const COLLECTION_ITEMS_RESOLVE_FAILED = 'COLLECTION_ITEMS_RESOLVE_FAILED';
 export const COLLECTION_NEW = 'COLLECTION_NEW';
-export const COLLECTION_RENAME = 'COLLECTION_RENAME';
 export const COLLECTION_DELETE = 'COLLECTION_DELETE';
 export const COLLECTION_PENDING = 'COLLECTION_PENDING';
 export const COLLECTION_EDIT = 'COLLECTION_EDIT';
