+++ ui/page/collection/index.js
@@ -54,6 +54,7 @@ const perform = (dispatch) => ({
   fetchCollectionItems: (claimId, cb) => dispatch(doFetchItemsInCollection({ collectionId: claimId }, cb)), // if this collection is not resolved, resolve it
   deleteCollection: (id, colKey) => dispatch(doCollectionDelete(id, colKey)),
   editCollection: (id, params) => dispatch(doCollectionEdit(id, params)),
+  renameCollection: (id, name) => dispatch(doCollectionEdit(id, { name })),
 });
 
 export default withRouter(connect(select, perform)(CollectionPage));
